+++ CHANGELOG.md
@@ -47,3 +47,11 @@
 ## [Unreleased]
 
 Versions of TeXLite in development.
+
+### v1.2.0 (2020-09-09)
+
+- Made both $$$ and $$ valid for equation enclosure.
+- Made equations ignore empty lines.
+- Added boolean `:nopagenumbers:` option to remove page numbers.
+- Added `:pagesize:` option for page sizes (e.g.: `letter`, `a4paper`).
+- Added boolean `:twocolumn:` option for the document layout.